fix(atlas-ui): moved core variable file one level up
