+++ packages/theming/atlas/src/theme/native/main.ts
@@ -1,1 +1,1 @@
-import {} from "../../themesource/atlas_ui_resources/native/core/variables";
+import {} from "../../themesource/atlas_ui_resources/native/variables";
