+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/buildingblocks/card.ts
@@ -1,5 +1,5 @@
 import { Platform } from "react-native";
-import { background, border, contrast, spacing } from "../core/variables";
+import { background, border, contrast, spacing } from "../variables";
 /*
 ==========================================================================
     Cards
