+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/base/spacing.ts
@@ -1,4 +1,4 @@
-import { spacing } from "../variables";
+import { spacing } from "../../variables";
 import { Helperclass } from "../../types/helperclass";
 /*
