+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/badge.ts
@@ -1,4 +1,4 @@
-import { badge } from "../variables";
+import { badge } from "../../variables";
 /*
 
 DISCLAIMER:
