+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/buttons.ts
@@ -1,4 +1,4 @@
-import { brand, button, contrast, font } from "../variables";
+import { brand, button, contrast, font } from "../../variables";
 import merge from "./_functions/mergeobjects";
 import { ActionButtonType } from "../../types/widgets";
 /*
