+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/floatingactionbutton.ts
@@ -1,4 +1,4 @@
-import { background, brand } from "../variables";
+import { background, brand } from "../../variables";
 /*
 
 DISCLAIMER:
