+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/images.ts
@@ -1,4 +1,4 @@
-import { border, brand, contrast, image } from "../variables";
+import { border, brand, contrast, image } from "../../variables";
 /*
 
 DISCLAIMER:
