+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/listviewswipe.ts
@@ -1,4 +1,4 @@
-import { listViewSwipe } from "../variables";
+import { listViewSwipe } from "../../variables";
 
 /*
