+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/maps.ts
@@ -1,4 +1,4 @@
-import { border, brand } from "../variables";
+import { border, brand } from "../../variables";
 /*
 
 DISCLAIMER:
