+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/progressbar.ts
@@ -1,4 +1,4 @@
-import { brand, progressBar } from "../variables";
+import { brand, progressBar } from "../../variables";
 /*
 
 DISCLAIMER:
