+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/slider.ts
@@ -1,4 +1,4 @@
-import { brand } from "../variables";
+import { brand } from "../../variables";
 /*
 
 DISCLAIMER:
