+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/helpers/typography.ts
@@ -1,4 +1,4 @@
-import { brand, font } from "../variables";
+import { brand, font } from "../../variables";
 /*
 
 DISCLAIMER:
