+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/activityindicator.ts
@@ -1,4 +1,4 @@
-import { brand } from "../variables";
+import { brand } from "../../variables";
 import { ActivityIndicatorType } from "../../types/widgets";
 /*
