+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/barchart.ts
@@ -1,4 +1,4 @@
-import { border, brand, font, spacing } from "../variables";
+import { border, brand, font, spacing } from "../../variables";
 import { BarChartType } from "../../types/widgets";
 /*
