+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/bottomsheet.ts
@@ -1,4 +1,4 @@
-import { background, border, brand, contrast, font } from "../variables";
+import { background, border, brand, contrast, font } from "../../variables";
 import { BottomSheetType } from "../../types/widgets";
 
 //
