+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/carousel.ts
@@ -1,4 +1,4 @@
-import { brand, contrast, font, spacing } from "../variables";
+import { brand, contrast, font, spacing } from "../../variables";
 import { CarouselLayoutType, CarouselType } from "../../types/widgets";
 /*
