+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/checkbox.ts
@@ -1,5 +1,5 @@
 import { Platform } from "react-native";
-import { background, border, brand, contrast, font, spacing } from "../variables";
+import { background, border, brand, contrast, font, spacing } from "../../variables";
 import { TextBox, TextBoxVertical } from "./textbox";
 import { CheckBoxType } from "../../types/widgets";
 /*
