+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/dropdown.ts
@@ -1,4 +1,4 @@
-import { font, input } from "../variables";
+import { font, input } from "../../variables";
 import { TextBox, TextBoxVertical } from "./textbox";
 import { DropDownType } from "../../types/widgets";
 /*
