+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/images.ts
@@ -1,4 +1,4 @@
-import { contrast, image } from "../variables";
+import { contrast, image } from "../../variables";
 import { ImageType } from "../../types/widgets";
 /*
