+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/listviewswipe.ts
@@ -1,4 +1,4 @@
-import { background, listViewSwipe } from "../variables";
+import { background, listViewSwipe } from "../../variables";
 import { ListViewSwipeType } from "../../types/widgets";
 /*
