+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/qrcode.ts
@@ -1,4 +1,4 @@
-import { background, contrast } from "../variables";
+import { background, contrast } from "../../variables";
 import { QRCodeType } from "../../types/widgets";
 /*
