+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/slider.ts
@@ -1,4 +1,4 @@
-import { slider } from "../variables";
+import { slider } from "../../variables";
 import { SliderType } from "../../types/widgets";
 import { TextBox } from "./textbox";
 /*
