+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/textbox.ts
@@ -1,4 +1,4 @@
-import { font, input, spacing } from "../variables";
+import { font, input, spacing } from "../../variables";
 import { TextBoxType } from "../../types/widgets";
 /*
