+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/togglebuttons.ts
@@ -1,4 +1,4 @@
-import { background, brand, contrast, font } from "../variables";
+import { background, brand, contrast, font } from "../../variables";
 import { ToggleButtonsType } from "../../types/widgets";
 import { TextBox } from "./textbox";
 /*
