+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/core/widgets/typography.ts
@@ -1,4 +1,4 @@
-import { font } from "../variables";
+import { font } from "../../variables";
 import { TextType } from "../../types/widgets";
 /*
