+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/layouts/layout.ts
@@ -1,6 +1,6 @@
 import { NativeModules } from "react-native";
 import { darkMode } from "../../../../theme/native/custom-variables";
-import { background, font, navigation } from "../core/variables";
+import { background, font, navigation } from "../variables";
 import { LayoutType } from "../types/widgets";
 /*
