+++ packages/theming/atlas/src/themesource/atlas_ui_resources/native/layouts/page.ts
@@ -1,4 +1,4 @@
-import { background, font, navigation } from "../core/variables";
+import { background, font, navigation } from "../variables";
 import { LayoutType } from "../types/widgets";
 /*
