+++ packages/theming/atlas/src/themesource/atlas_ui_resources/web/_variables.scss
