+++ packages/theming/atlas/src/themesource/atlas_ui_resources/web/main.scss
@@ -6,7 +6,7 @@
 //================================== CORE ==================================\\
 
 // Default variables
-@import "core/variables";
+@import "variables";
 
 // Base
 @import "core/base/mixins/animations";
